requirements.txt
psycopg2-binary
sqlalchemy
polars
python-dotenv
requests